Correccion imagenes y fecha set automatica

+++ ventas/src/main/java/com/example/ventas/Models/Cliente.java
@@ -31,7 +31,7 @@ public class Cliente {
     @Column
     private String direccion;
     @Column
-    private int telefono;
+    private Long telefono;
